\name{glam-package}
\alias{glam-package}
\alias{glam}
\docType{package}
\title{
\packageTitle{glam}
}
\description{
\packageDescription{glam}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{glam}
\packageIndices{glam}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{glam}

Maintainer: \packageMaintainer{glam}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
